Auto tag on cos/system bump
This action will check after every master merge, check if the diff
contains changes to the cos package files, extract the current tag for
cos/system, get the latest published tag and compare both. If the
current cos/system version is different from the last tag, then it will
create a new tag with that version, pointing to the current commit.

Signed-off-by: Itxaka <igarcia@suse.com>